Implement a smarter HostPort object
Now HostPort will try to find actual localhost address AND a free
listening port (race condition permitting).

This means that the test port can no longer be expected to be
"localhost", which means that on *nix systems trying to connect to
"localhost" will likely fail (because "localhost" is invariably mapped
to "127.0.0.1")

So, a get_sock() classmethod contextmanager is implemented to assist
test cases. This affects test_proxyprotocol.py (previously hardcoded to
AF_INET) and test_smtp (previously hardcoded to default, which is
AF_INET).